IOTMS/M1/requirements.txt
paho-mqtt==1.6.1
pymongo==4.3.3
pytz